[PVR] Refactor CPVRGUIActions step 5: Move channel related actions to own class.
