+++ xbmc/pvr/guilib/PVRGUIActions.h
@@ -8,18 +8,16 @@
 
 #pragma once
 
-#include "pvr/PVRChannelNumberInputHandler.h"
+#include "pvr/guilib/PVRGUIActionsChannels.h"
 #include "pvr/guilib/PVRGUIActionsEPG.h"
 #include "pvr/guilib/PVRGUIActionsPlayback.h"
 #include "pvr/guilib/PVRGUIActionsRecordings.h"
 #include "pvr/guilib/PVRGUIActionsTimers.h"
-#include "pvr/guilib/PVRGUIChannelNavigator.h"
 #include "pvr/settings/PVRSettings.h"
 #include "threads/CriticalSection.h"
 
 #include <memory>
 #include <string>
-#include <vector>
 
 class CFileItem;
 
@@ -33,31 +31,10 @@ enum class ParentalCheckResult
 };
 
 class CPVRChannel;
-class CPVRChannelGroupMember;
 class CPVRTimerInfoTag;
 
-class CPVRChannelSwitchingInputHandler : public CPVRChannelNumberInputHandler
-{
-public:
-  // CPVRChannelNumberInputHandler implementation
-  void GetChannelNumbers(std::vector<std::string>& channelNumbers) override;
-  void AppendChannelNumberCharacter(char cCharacter) override;
-  void OnInputDone() override;
-
-private:
-  /*!
-   * @brief Switch to the channel with the given number.
-   * @param channelNumber the channel number
-   */
-  void SwitchToChannel(const CPVRChannelNumber& channelNumber);
-
-  /*!
-   * @brief Switch to the previously played channel.
-   */
-  void SwitchToPreviousChannel();
-};
-
-class CPVRGUIActions : public CPVRGUIActionsEPG,
+class CPVRGUIActions : public CPVRGUIActionsChannels,
+                       public CPVRGUIActionsEPG,
                        public CPVRGUIActionsPlayback,
                        public CPVRGUIActionsRecordings,
                        public CPVRGUIActionsTimers
@@ -65,31 +42,6 @@ class CPVRGUIActions : public CPVRGUIActionsEPG,
 public:
   CPVRGUIActions();
   virtual ~CPVRGUIActions() = default;
-
-  /*!
-   * @brief Hide a channel, always showing a confirmation dialog.
-   * @param item containing a channel or an epg tag.
-   * @return true on success, false otherwise.
-   */
-  bool HideChannel(const std::shared_ptr<CFileItem>& item) const;
-
-  /*!
-   * @brief Open a selection dialog and start a channel scan on the selected client.
-   * @return true on success, false otherwise.
-   */
-  bool StartChannelScan();
-
-  /*!
-   * @brief Start a channel scan on the specified client or open a dialog to select a client
-   * @param clientId the id of client to scan or PVR_INVALID_CLIENT_ID if a dialog will be opened
-   * @return true on success, false otherwise.
-   */
-  bool StartChannelScan(int clientId);
-
-  /*!
-   * @return True when a channel scan is currently running, false otherwise.
-   */
-  bool IsRunningChannelScan() const { return m_bChannelScanRunning; }
 
   /*!
    * @brief Select and invoke client-specific settings actions
@@ -144,47 +96,6 @@ public:
   void SetSelectedItemPath(bool bRadio, const std::string& path);
 
   /*!
-   * @brief Get a channel group member for the given channel, either from the currently active
-   * group or if not found there, from the 'all channels' group.
-   * @param channel the channel.
-   * @return the group member or nullptr if not found.
-   */
-  std::shared_ptr<CPVRChannelGroupMember> GetChannelGroupMember(
-      const std::shared_ptr<CPVRChannel>& channel) const;
-
-  /*!
-   * @brief Get a channel group member for the given item, either from the currently active group
-   * or if not found there, from the 'all channels' group.
-   * @param item the item containing a channel, channel group, recording, timer or epg tag.
-   * @return the group member or nullptr if not found.
-   */
-  std::shared_ptr<CPVRChannelGroupMember> GetChannelGroupMember(const CFileItem& item) const;
-
-  /*!
-   * @brief Get the currently active channel number input handler.
-   * @return the handler.
-   */
-  CPVRChannelNumberInputHandler& GetChannelNumberInputHandler();
-
-  /*!
-   * @brief Get the channel navigator.
-   * @return the navigator.
-   */
-  CPVRGUIChannelNavigator& GetChannelNavigator();
-
-  /*!
-   * @brief Inform GUI actions that playback of an item just started.
-   * @param item The item that started to play.
-   */
-  void OnPlaybackStarted(const std::shared_ptr<CFileItem>& item);
-
-  /*!
-   * @brief Inform GUI actions that playback of an item was stopped due to user interaction.
-   * @param item The item that stopped to play.
-   */
-  void OnPlaybackStopped(const std::shared_ptr<CFileItem>& item);
-
-  /*!
    * @brief Process info action for the given item.
    * @param item The item.
    */
@@ -199,10 +110,7 @@ private:
   bool IsNextEventWithinBackendIdleTime() const;
 
   mutable CCriticalSection m_critSection;
-  CPVRChannelSwitchingInputHandler m_channelNumberInputHandler;
-  bool m_bChannelScanRunning = false;
   CPVRSettings m_settings;
-  CPVRGUIChannelNavigator m_channelNavigator;
   std::string m_selectedItemPathTV;
   std::string m_selectedItemPathRadio;
 };
